Trait bitintr::Blsmsk

pub trait Blsmsk {
    fn blsmsk(self) -> Self;
}

Get mask up to lowest set bit.

Required methods

fn blsmsk(self) -> Self

Get mask up to lowest set bit.

Sets all the bits of the result to 1 up to and including the lowest set bit of self.

If self is zero, all the bits of the result are set.

Instructions

BLSMSK:

Description: Get mask up to lowest set bit.

Architecture: x86.

Instruction set: BMI.

Registers: 32/64 bit.

Example

assert_eq!(0b0011_0000u8.blsmsk(), 0b0001_1111u8);
assert_eq!(0b0000_0000u8.blsmsk(), 0b1111_1111u8);

Loading content...