Trait bitintr::T1mskc

pub trait T1mskc {
    fn t1mskc(self) -> Self;
}

Inverse mask from trailing ones

Required methods

fn t1mskc(self) -> Self

Clears all bits below the least significant zero of self and sets all other bits.

If the least significant bit of self is 0, it sets all bits.

Instructions

T1MSKC:

Description: Inverse mask from trailing ones

Architecture: x86.

Instruction set: TBM.

Registers: 32/64 bit.

Example

assert_eq!(0b0101_0111u8.t1mskc(), 0b1111_1000u8);
assert_eq!(0b0101_0110u8.t1mskc(), 0b1111_1111u8);

Loading content...