Function bitintr::x86::abm::lzcnt

pub fn lzcnt<T: Int>(x: T) -> T

Counts the leading most significant zero bits.

When the operand is zero, it returns its size in bits.

Keywords: Leading Zeros Count, Count Leading Zeros, Bit Scan Severse, Find Last Set.

See also arm::v7::clz.

Assembly Instructions

LZCNT:

Description: Count the number of leading zero bits.

Architecture: x86.

Instruction set: ABM, BMI.

Registers: 16/32/64 bit.

Note: This instruction is officially part of BMI1 but Intel (and AMD) CPUs advertise it as being part of ABM.

CLZ:

Description: Count Leading Zeros.

Architecture: ARM.

Instruction set: v7.

Registers: 8/16/32 bit.

Example

use bitintr::x86::abm::*;

assert_eq!(lzcnt(0b0101_1010u16), 9u16);
assert_eq!(0b0101_1010u16.lzcnt(), 9u16);