Module bitintr::x86::bmi

Bit Manipulation Instruction (BMI) Set 1.0.

The reference is Intel 64 and IA-32 Architectures Software Developer's Manual Volume 2: Instruction Set Reference, A-Z.

Wikipedia provides a quick overview of the available instructions.

Traits

ANDN
Method version of andn.

BEXTR
Method version of bextr.

BLSI
Method version of blsi.

BLSMSK
Method version of blsmsk.

BLSR
Method version of blsr.

TZCNT
Method version of tzcnt.

Functions

andn
Bitwise logical AND of inverted x with y.

bextr
Bit Field Extract.

blsi
Extract lowest set isolated bit.

blsmsk
Get mask up to lowest set bit.

blsr
Resets the lowest set bit of x.

tzcnt
Counts the number of trailing least significant zero bits.