Module bitintr::x86::bmi2

Bit Manipulation Instruction (BMI) Set 2.0.

The reference is Intel 64 and IA-32 Architectures Software Developer's Manual Volume 2: Instruction Set Reference, A-Z.

Wikipedia provides a quick overview of the available instructions.

Traits

BZHI
Method version of bzhi.

MULX
Method version of mulx.

PDEP
Method version of pdep.

PEXT
Method version of pext.

Functions

bzhi
Zero the high bits of x at position >= bit_position.

mulx
Unsigned multiply without affecting flags.

pdep
Parallel bits deposit.

pext
Parallel bits extract.