Module bitintr::x86::tbm

Trailing Bit Manipulation (TBM) instruction set.

The reference is AMD64 Architecture Programmer's Manual, Volume 3: General-Purpose and System Instructions.

Wikipedia provides a quick overview of the available instructions.

Traits

BEXTR
Method version of bextr.

BLCFILL
Method version of blcfill.

BLCI
Method version of blci.

BLCIC
Method version of blcic.

BLCMSK
Method version of blcmsk.

BLCS
Method version of blcs.

BLSFILL
Method version of blsfill.

BLSIC
Method version of blsic.

T1MSKC
Method version of t1mskc.

TZMSK
Method version of tzmsk.

Functions

bextr
Bit Field Extract.

blcfill
Clears all bits below the least significant zero bit of x.

blci
Sets all bits of x to 1 except for the least significant zero bit.

blcic
Sets the least significant bit of x and clears all other bits.

blcmsk
Sets the least significant bit of x and clears all bits above that bit.

blcs
Sets the least significant bit of x.

blsfill
Sets all bits of x below the least significant one.

blsic
Clears least significant bit and sets all other bits.

t1mskc
Clears all bits below the least significant zero of x and sets all other bits.

tzmsk
Sets all bits below the least significant one of x and clears all other bits.