Function bitintr::x86::tbm::blcic

pub fn blcic<T: Int>(x: T) -> T

Sets the least significant bit of x and clears all other bits.

If there is no zero bit in x, it returns zero.

Intrinsic (when available TBM)

BLCIC: Isolate lowest clear bit (supports 32/64 bit registers).

Example

use bitintr::x86::tbm::*;

assert_eq!(blcic(0b0101_0001u8), 0b0000_0010u8);
assert_eq!(blcic(0b1111_1111u8), 0b0000_0000u8);